refactor(grimoire): Phase 2 - Extract TechParticles to modular structure
Enterprise-level code organization following BookScene/ConstellationLine patterns:

New Structure:
- utils/math.ts: distance calculation helper
- utils/starTexture.ts: star texture generation (83 lines)
- components/scene/TechIcon.tsx: icon component (23 lines)
- TechParticles.tsx: reduced from 452 to ~350 lines

Changes:
- Extracted 70-line createStarTexture function to dedicated utility
- Separated TechIcon sub-component to own file
- Moved distance helper to math utilities
- Fixed TechIcon size from [2,2] to [1,1]
- Removed unused THREE import from math.ts

Benefits:
- Single responsibility per file
- Improved testability and maintainability
- Matches enterprise patterns of other components
- Cleaner imports and dependencies
