+++ src/features/forge/home/components/tech-grimoire/utils/math.ts
@@ -1,0 +1,6 @@
+export const distance = (
+  p1: { x: number; y: number; z: number },
+  p2: { x: number; y: number; z: number }
+): number => {
+  return Math.sqrt(Math.pow(p2.x - p1.x, 2) + Math.pow(p2.y - p1.y, 2) + Math.pow(p2.z - p1.z, 2));
+};
